refactor: Move files to folders created on develop branch to make merging easier. Adjust Makefile and .gitignore.
