+++ Makefile
@@ -6,7 +6,7 @@
 #    By: gmachado <gmachado@student.42sp.org.br>    +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/03/19 16:33:22 by macarval          #+#    #+#              #
-#    Updated: 2024/06/18 09:06:05 by macarval         ###   ########.fr        #
+#    Updated: 2024/06/25 04:49:54 by gmachado         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -15,9 +15,10 @@ NAME		= ircserv
 SRCS		= main.cpp Bot.cpp Channel.cpp ChannelList.cpp Client.cpp \
 			ClientList.cpp Commands.cpp FileTransfer.cpp IrcServer.cpp
 
-VPATH		= src/
+VPATH		= src/ src/bonus src/channel src/client src/commands
 OBJS_PATH	= obj
-INCLUDE		= -I./include
+INCLUDE		= -I./include -I./include/bonus -I./include/channel \
+			-I./include/client -I./include/commands
 
 FLAGS		= -g3 -Wall -Wextra -Werror -g -std=c++98 -Wpedantic
 CC			= c++
