--- include/colors.hpp
@@ -1,28 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   colors.hpp                                         :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: gmachado <gmachado@student.42sp.org.br>    +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/06/18 03:05:41 by gmachado          #+#    #+#             */
-/*   Updated: 2024/06/18 03:07:00 by gmachado         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#ifndef COLORS_HPP
-# define COLORS_HPP
-
-#include <string>
-
-// Color codes as global variables
-const std::string RESET = "\033[0m";
-const std::string RED = "\033[31;1m";
-const std::string GREEN = "\033[32;1m";
-const std::string YELLOW = "\033[33;1m";
-const std::string BLUE = "\033[34;1m";
-const std::string PURPLE = "\033[35;1m";
-const std::string CYAN = "\033[36;1m";
-const std::string GRAY = "\033[37;1m";
-
-#endif
