+++ src/bonus/Bot.cpp
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../include/Bot.hpp"
+#include "Bot.hpp"
 
 // Constructor ================================================================
 Bot::Bot() {}
