+++ src/channel/Channel.cpp
